final class instead of interface
